+++ services/slack/service.go
@@ -2,6 +2,8 @@ package slack
 
 import (
 	"bytes"
+	"crypto/tls"
+	"crypto/x509"
 	"encoding/json"
 	"fmt"
 	"io"
@@ -16,15 +18,29 @@ import (
 
 type Service struct {
 	configValue atomic.Value
+	clientValue atomic.Value
 	logger      *log.Logger
+	client      *http.Client
 }
 
-func NewService(c Config, l *log.Logger) *Service {
+func NewService(c Config, l *log.Logger) (*Service, error) {
+	tlsConfig, err := getTLSConfig(c.SSLCA, c.SSLCert, c.SSLKey, c.InsecureSkipVerify)
+	if err != nil {
+		return nil, err
+	}
+	if tlsConfig.InsecureSkipVerify {
+		l.Println("W! Slack service is configured to skip ssl verification")
+	}
 	s := &Service{
 		logger: l,
 	}
 	s.configValue.Store(c)
-	return s
+	s.clientValue.Store(&http.Client{
+		Transport: &http.Transport{
+			TLSClientConfig: tlsConfig,
+		},
+	})
+	return s, nil
 }
 
 func (s *Service) Open() error {
@@ -46,7 +62,19 @@ func (s *Service) Update(newConfig []interface{}) error {
 	if c, ok := newConfig[0].(Config); !ok {
 		return fmt.Errorf("expected config object to be of type %T, got %T", c, newConfig[0])
 	} else {
+		tlsConfig, err := getTLSConfig(c.SSLCA, c.SSLCert, c.SSLKey, c.InsecureSkipVerify)
+		if err != nil {
+			return err
+		}
+		if tlsConfig.InsecureSkipVerify {
+			s.logger.Println("W! Slack service is configured to skip ssl verification")
+		}
 		s.configValue.Store(c)
+		s.clientValue.Store(&http.Client{
+			Transport: &http.Transport{
+				TLSClientConfig: tlsConfig,
+			},
+		})
 	}
 	return nil
 }
@@ -99,7 +127,8 @@ func (s *Service) Alert(channel, message, username, iconEmoji string, level aler
 	if err != nil {
 		return err
 	}
-	resp, err := http.Post(url, "application/json", post)
+	client := s.clientValue.Load().(*http.Client)
+	resp, err := client.Post(url, "application/json", post)
 	if err != nil {
 		return err
 	}
@@ -209,4 +238,40 @@ func (h *handler) Handle(event alert.Event) {
 	); err != nil {
 		h.logger.Println("E! failed to send event to Slack", err)
 	}
+}
+
+// getTLSConfig creates a tls.Config object from the given certs, key, and CA files.
+// you must give the full path to the files.
+func getTLSConfig(
+	SSLCA, SSLCert, SSLKey string,
+	InsecureSkipVerify bool,
+) (*tls.Config, error) {
+	t := &tls.Config{
+		InsecureSkipVerify: InsecureSkipVerify,
+	}
+	if SSLCert != "" && SSLKey != "" {
+		cert, err := tls.LoadX509KeyPair(SSLCert, SSLKey)
+		if err != nil {
+			return nil, fmt.Errorf(
+				"Could not load TLS client key/certificate: %s",
+				err)
+		}
+		t.Certificates = []tls.Certificate{cert}
+	} else if SSLCert != "" {
+		return nil, errors.New("Must provide both key and cert files: only cert file provided.")
+	} else if SSLKey != "" {
+		return nil, errors.New("Must provide both key and cert files: only key file provided.")
+	}
+
+	if SSLCA != "" {
+		caCert, err := ioutil.ReadFile(SSLCA)
+		if err != nil {
+			return nil, fmt.Errorf("Could not load TLS CA: %s",
+				err)
+		}
+		caCertPool := x509.NewCertPool()
+		caCertPool.AppendCertsFromPEM(caCert)
+		t.RootCAs = caCertPool
+	}
+	return t, nil
 }